% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cisLeave1out.R
\name{cisLeave1out}
\alias{cisLeave1out}
\title{cisLeave1out}
\usage{
cisLeave1out(
  SNP,
  beta.exposure,
  se.exposure,
  beta.outcome,
  se.outcome,
  method = "RE",
  sm,
  gene = NA,
  matrix
)
}
\arguments{
\item{method}{fixed (FE) or random effects (RE, default) IVW}

\item{sm}{Estimator type (see meta R package documentation)}

\item{gene}{colname specifying which gene SNPs come from if stratifying by locus will be useful.}

\item{matrix}{SNP correlation matrix}

\item{betaXG}{SNP-exposure associations.}

\item{betaYG}{SNP-outcome associations.}

\item{sebetaXG}{Standard error of the SNP-exposure associations.}

\item{sebetaYG}{Standard error of the SNP-outcome associations}

\item{FirstOrderSE}{Use first or second order approximations of the standard error. Default is to use first order (i.e. SE = sebetaYG/betaXG)}
}
\description{
This function implements a leave one out analysis using IVW with correlated variants. Requires installation of the meta R package.
}
