% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGGERcorrel.R
\name{EGGERcorrel}
\alias{EGGERcorrel}
\title{EGGERcorrel}
\usage{
EGGERcorrel(
  betaYG,
  sebetaYG,
  betaXG,
  sebetaXG,
  rho,
  method = "GWLR",
  FirstOrderSE = T
)
}
\arguments{
\item{betaYG}{SNP-outcome associations.}

\item{sebetaYG}{Standard error of the SNP-outcome associations}

\item{betaXG}{SNP-exposure associations.}

\item{sebetaXG}{Standard error of the SNP-exposure associations.}

\item{rho}{SNP correlation matrix}

\item{method}{method for accounting for SNP correlations. Can be set to "GWLR" (generalized weighted linear regression, the default) or "LRwCd"  (linear regression command after weighting the data by a Cholesky decomposition)}

\item{FirstOrderSE}{Use first or second order approximations of the standard error. Default is to use first order (i.e. SE = sebetaYG/betaXG)}
}
\description{
This function implements MR Egger with correlated variants. Code taken from https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5506233/
}
