% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_cor.R
\name{F_cor}
\alias{F_cor}
\title{F_cor}
\usage{
F_cor(betaXG, sebetaXG, rho, N = NA)
}
\arguments{
\item{betaXG}{SNP-exposure associations.}

\item{sebetaXG}{Standard error of the SNP-exposure associations.}

\item{rho}{SNP correlation matrix}

\item{N}{The sample size of the GWAS}
}
\description{
This function estimates the F stat with correlated variants. If N is specified then it will also return the r^2
}
