% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVWcorrel.R
\name{IVWcorrel}
\alias{IVWcorrel}
\title{IVWcorrel}
\usage{
IVWcorrel(
  betaYG,
  sebetaYG,
  betaXG,
  sebetaXG,
  rho,
  method = "RE",
  FirstOrderSE = T
)
}
\arguments{
\item{betaYG}{SNP-outcome associations.}

\item{sebetaYG}{Standard error of the SNP-outcome associations}

\item{betaXG}{SNP-exposure associations.}

\item{sebetaXG}{Standard error of the SNP-exposure associations.}

\item{rho}{SNP correlation matrix}

\item{method}{fixed (FE) or random effects (RE, default)}

\item{FirstOrderSE}{Use first or second order approximations of the standard error. Default is to use first order (i.e. SE = sebetaYG/betaXG)}
}
\description{
This function implements IVW with correlated variants. Code taken from DOI: 10.1002/gepi.22077
}
