% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSCMR.R
\name{TSCMR}
\alias{TSCMR}
\title{Two-Step Cis MR: A novel method and R package for attenuating bias in cis-MR in the absence of uncorrelated variants}
\usage{
TSCMR(Bgo, SEgo, Bgc, SEgc, Bco, SEco, straps = 1e+06)
}
\arguments{
\item{Bgo}{SNP-outcome associations.}

\item{SEgo}{Standard error of the SNP-outcome associations.}

\item{Bgc}{SNP-confounder associations.}

\item{SEgc}{Standard error of the SNP-confounder associations.}

\item{Bco}{confounder-outcome association.}

\item{SEco}{Standard error of the confounder-outcome association.}

\item{straps}{Number of iterations for the bootstrap standard error}
}
\description{
This function implements a simple adjsutment to the snp-outcome association to remove bias from cis-MR estimates in the absence of uncorrelated variants.
}
\keyword{2SMR}
