% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStepCisMR-package.R
\docType{package}
\name{TwoStepCisMR-package}
\alias{TwoStepCisMR}
\alias{TwoStepCisMR-package}
\title{Two-Step Cis MR: A novel method and R package for attenuating bias in cis-MR in the absence of uncorrelated variants.}
\description{
Two-Step Cis MR: A novel method and R package for attenuating bias in cis-MR in the absence of uncorrelated variants
}
\author{
Benji Woolf, benjamin.woolf@bristol.ac.uk
}
\keyword{2SMR}
